agent-swarm/requirements.txt
openhands-ai
pydantic

